Extract CIAB plan upgrade URL building to CIABSiteGateKeeper
